Refactor profile methods
