+++ code/api/imports/data/collection/methods/Profile/fetchOneProfile.js
@@ -1,0 +1,6 @@
+import { profileCollection } from '../../ProfileCollection'
+
+export const fetchOneProfile = referenceId => profileCollection.findOne({
+  referenceId,
+  type: 'user', // FIXME type is not needed
+})
